Clean up includes, move some functions to IMesh file
Move mesh functions to IMesh file as they shouldn't need knowledge of Mesh
#741
