+++ trview.app/Geometry/Mesh.h
@@ -58,132 +58,4 @@ namespace trview
         std::vector<Triangle>                             _collision_triangles;
         DirectX::BoundingBox                              _bounding_box;
     };
-
-    /// Create a new mesh based on the contents of the mesh specified.
-    /// @param level_version The level version - affects texture index
-    /// @param mesh The level mesh to generate.
-    /// @param source The mesh source function.
-    /// @param texture_storage The textures for the level.
-    /// @param transparent_collision Whether to include transparent triangles in collision triangles.
-    /// @returns The new mesh.
-    std::unique_ptr<IMesh> create_mesh(trlevel::LevelVersion level_version, const trlevel::tr_mesh& mesh, const IMesh::Source& source, const ILevelTextureStorage& texture_storage, bool transparent_collision = true);
-
-    /// Create a new cube mesh.
-    std::unique_ptr<IMesh> create_cube_mesh(const IMesh::Source& source);
-
-    enum class SpriteOffsetMode
-    {
-        RoomSprite,
-        Entity
-    };
-
-    /// Create a new sprite mesh.
-    /// @param source The function to call to create a IMesh
-    /// @param sprite The game sprite definition to use.
-    /// @param scale The output scale matrix.
-    /// @param offset The output offset.
-    std::unique_ptr<IMesh> create_sprite_mesh(
-        const IMesh::Source& source,
-        const trlevel::tr_sprite_texture& sprite,
-        DirectX::SimpleMath::Matrix& scale,
-        DirectX::SimpleMath::Vector3& offset,
-        SpriteOffsetMode offset_mode);
-
-    /// Create a new sprite mesh.
-    /// @param source The function to call to create a IMesh
-    /// @param sprite The game sprite definition to use.
-    /// @param scale The output scale matrix.
-    std::unique_ptr<IMesh> create_sprite_mesh(
-        const IMesh::Source& source,
-        const trlevel::tr_sprite_texture& sprite,
-        DirectX::SimpleMath::Matrix& scale,
-        DirectX::SimpleMath::Matrix& offset,
-        SpriteOffsetMode offset_mode);
-
-    /// Convert the textured rectangles into collections required to create a mesh.
-    /// @param level_version The level version - affects texture index.
-    /// @param rectangles The rectangles from the mesh or room geometry.
-    /// @param input_vertices The vertices that the rectangle indices refer to.
-    /// @param texture_storage The texture storage for the level.
-    /// @param output_vertices The collection to add the new vertices to.
-    /// @param output_indices The collection to add new indices to.
-    /// @param transparent_triangles The collection to add transparent triangles to.
-    /// @param collision_triangles The collection to add collision triangles to.
-    /// @param transparent_collision Whether to add transparent rectangles as collision triangles.
-    void process_textured_rectangles(
-        trlevel::LevelVersion level_version,
-        const std::vector<trlevel::tr4_mesh_face4>& rectangles, 
-        const std::vector<trlevel::tr_vertex>& input_vertices, 
-        const ILevelTextureStorage& texture_storage,
-        std::vector<MeshVertex>& output_vertices,
-        std::vector<std::vector<uint32_t>>& output_indices,
-        std::vector<TransparentTriangle>& transparent_triangles,
-        std::vector<Triangle>& collision_triangles,
-        bool transparent_collision = true);
-
-    /// Convert the textured triangles into collections required to create a mesh.
-    /// @param triangles The triangles from the mesh or room geometry.
-    /// @param input_vertices The vertices that the triangle indices refer to.
-    /// @param texture_storage The texture storage for the level.
-    /// @param output_vertices The collection to add the new vertices to.
-    /// @param output_indices The collection to add new indices to.
-    /// @param transparent_triangles The collection to add transparent triangles to.
-    /// @param collision_triangles The collection to add collision triangles to.
-    /// @param transparent_collision Whether to add transparent rectangles as collision triangles.
-    void process_textured_triangles(
-        trlevel::LevelVersion level_version,
-        const std::vector<trlevel::tr4_mesh_face3>& triangles,
-        const std::vector<trlevel::tr_vertex>& input_vertices,
-        const ILevelTextureStorage& texture_storage,
-        std::vector<MeshVertex>& output_vertices,
-        std::vector<std::vector<uint32_t>>& output_indices,
-        std::vector<TransparentTriangle>& transparent_triangles,
-        std::vector<Triangle>& collision_triangles,
-        bool transparent_collision = true);
-
-    // Convert the coloured rectangles into collections required to create a mesh.
-    // triangles: The rectangles from the mesh or room geometry.
-    // input_vertices: The vertices that the rectangle indices refer to.
-    // texture_storage: The texture storage for the level.
-    // output_vertices: The collection to add the new vertices to.
-    // output_indices: The collection to add new indices to.
-    // collision_triangles: The collection to add collision triangles to.
-    void process_coloured_rectangles(
-        const std::vector<trlevel::tr_face4>& rectangles,
-        const std::vector<trlevel::tr_vertex>& input_vertices,
-        const ILevelTextureStorage& texture_storage,
-        std::vector<MeshVertex>& output_vertices,
-        std::vector<uint32_t>& output_indices,
-        std::vector<Triangle>& collision_triangles);
-
-    // Convert the coloured triangles into collections required to create a mesh.
-    // triangles: The triangles from the mesh or room geometry.
-    // input_vertices: The vertices that the triangle indices refer to.
-    // texture_storage: The texture storage for the level.
-    // output_vertices: The collection to add the new vertices to.
-    // output_indices: The collection to add new indices to.
-    // collision_triangles: The collection to add collision triangles to.
-    void process_coloured_triangles(
-        const std::vector<trlevel::tr_face3>& triangles,
-        const std::vector<trlevel::tr_vertex>& input_vertices,
-        const ILevelTextureStorage& texture_storage,
-        std::vector<MeshVertex>& output_vertices,
-        std::vector<uint32_t>& output_indices,
-        std::vector<Triangle>& collision_triangles);
-
-    /// Convert the vertex to the scale used by the viewer.
-    /// @param vertex The vertex to convert.
-    /// @retrurns The scaled vector.
-    DirectX::SimpleMath::Vector3 convert_vertex(const trlevel::tr_vertex& vertex);
-
-#pragma warning(push)
-#pragma warning(disable : 4324)
-    __declspec(align(16)) struct MeshData
-    {
-        DirectX::SimpleMath::Matrix matrix;
-        DirectX::SimpleMath::Color colour;
-        DirectX::SimpleMath::Vector4 light_dir;
-        int light_enabled{ 0 };
-    };
-#pragma warning(pop)
 }
