+++ trview.app/Tools/Measure.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <optional>
-#include <trview.app/Geometry/Mesh.h>
+#include <trview.app/Geometry/IMesh.h>
 #include "IMeasure.h"
 
 namespace trview
